Fix express router var name in example.
Part way down The router var magically changes from the declared `app` to an undeclared `server`, and then back again. Copy-paste error or partially-complete refactoring, probably? This won't compile in a lot of environments, and will definitely crash if you made it as far as running it.

I realize it's not actually meant to be run, but the example code should at least be _correct_, if not in fact _useful_.